+++ com.microej.example.iot.ssl.mqtt.publisher/src/main/java/com/microej/example/iot/mqtt/app/PublisherBG.java
@@ -1,46 +1,31 @@
 /*
  * Java
  *
- * Copyright 2016 IS2T. All rights reserved.
- * Use of this source code is subject to license terms.
+ * Copyright 2016-2018 IS2T. All rights reserved.
+ * For demonstration purpose only.
+ * IS2T PROPRIETARY. Use is subject to license terms.
  */
 package com.microej.example.iot.mqtt.app;
 
 import com.microej.example.iot.mqtt.HelloWorldPublisher;
+
 import ej.wadapps.app.BackgroundService;
 /**
  *
  */
-public class PublisherBG extends Thread implements BackgroundService {
+public class PublisherBG implements BackgroundService {
 
-	private static final int DELAY = 5000; // 5s
-	private boolean running = false;
+	private HelloWorldPublisher helloWorldPublisher;
 
 	@Override
 	public void onStart() {
-		running = true;
-		this.start();
-
+		helloWorldPublisher = new HelloWorldPublisher();
+		helloWorldPublisher.start();
 	}
 
 	@Override
 	public void onStop() {
-		running = false;
-		this.interrupt();
-
+		helloWorldPublisher.stop();
+		helloWorldPublisher = null;
 	}
-
-	@Override
-	public void run() {
-		while (running) {
-			HelloWorldPublisher.main(new String[0]);
-			try {
-				Thread.sleep(DELAY);
-			} catch (InterruptedException e) {
-				e.printStackTrace();
-			}
-		}
-	}
-
-
 }