+++ com.microej.example.iot.ssl.rest/src/main/java/com/microej/example/iot/ssl/rest/PostContent.java
@@ -1,9 +1,9 @@
 /*
  * Java
  *
- * Copyright 2015-2016 IS2T. All rights reserved.
- * Use of this source code is subject to license terms.
- *
+ * Copyright 2015-2018 IS2T. All rights reserved.
+ * For demonstration purpose only.
+ * IS2T PROPRIETARY. Use is subject to license terms.
  */
 package com.microej.example.iot.ssl.rest;
 
@@ -19,8 +19,9 @@ public class PostContent extends Content{
 		super(aMimeType, someBytes);
 	}
 
+	@Override
 	protected void addContent(URLConnection con) throws IOException{
-		((HttpURLConnection) con).setRequestMethod("POST");
+		((HttpURLConnection) con).setRequestMethod("POST"); //$NON-NLS-1$
 		super.addContent(con);
 	}
 }